Bug fix from PR-review
