+++ tickets/src/main/java/no/unit/nva/publication/ticket/update/UpdateTicketAssigneeHandler.java
@@ -1,0 +1,62 @@
+package no.unit.nva.publication.ticket.update;
+
+import com.amazonaws.services.lambda.runtime.Context;
+import java.net.HttpURLConnection;
+import no.unit.nva.publication.model.business.TicketEntry;
+import no.unit.nva.publication.service.impl.TicketService;
+import no.unit.nva.publication.ticket.TicketDto;
+import no.unit.nva.publication.ticket.TicketHandler;
+import nva.commons.apigateway.AccessRight;
+import nva.commons.apigateway.RequestInfo;
+import nva.commons.apigateway.exceptions.ApiGatewayException;
+import nva.commons.apigateway.exceptions.ForbiddenException;
+import nva.commons.apigateway.exceptions.UnauthorizedException;
+import nva.commons.core.JacocoGenerated;
+
+public class UpdateTicketAssigneeHandler extends TicketHandler<TicketDto, Void> {
+
+    private final TicketService ticketService;
+
+    @JacocoGenerated
+    public UpdateTicketAssigneeHandler() {
+        this(TicketService.defaultService()
+        );
+    }
+
+    public UpdateTicketAssigneeHandler(TicketService ticketService) {
+        super(TicketDto.class);
+        this.ticketService = ticketService;
+    }
+
+    @Override
+    protected Void processInput(TicketDto input, RequestInfo requestInfo, Context context) throws ApiGatewayException {
+        var ticketIdentifier = extractTicketIdentifierFromPath(requestInfo);
+        var ticket = ticketService.fetchTicketByIdentifier(ticketIdentifier);
+
+        if (userIsNotAuthorized(requestInfo, ticket)) {
+            throw new ForbiddenException();
+        }
+
+        ticketService.updateTicketAssignee(ticket, input.getAssignee());
+        return null;
+    }
+
+    @Override
+    protected Integer getSuccessStatusCode(TicketDto input, Void output) {
+        return HttpURLConnection.HTTP_ACCEPTED;
+    }
+
+    private static boolean userIsNotAuthorized(RequestInfo requestInfo, TicketEntry ticket)
+        throws UnauthorizedException {
+        return !(isAuthorizedToClaimTickets(requestInfo) && isUserFromSameCustomerAsTicket(requestInfo, ticket));
+    }
+
+    private static boolean isUserFromSameCustomerAsTicket(RequestInfo requestInfo, TicketEntry ticket)
+        throws UnauthorizedException {
+        return requestInfo.getCurrentCustomer().equals(ticket.getCustomerId());
+    }
+
+    private static boolean isAuthorizedToClaimTickets(RequestInfo requestInfo) {
+        return requestInfo.userIsAuthorized(AccessRight.APPROVE_DOI_REQUEST.toString());
+    }
+}
